Terragrunt should be able to pull in modules outside the current subfolders
When applying changes in a one environment (e.g. stage), the modules in
that environment may depend on modules in another environment (e.g.
mgmt). Update Terragrunt to be able to handle such “external”
dependencies, prompting the user whether it should assume that those
external dependencies are already applied (the default) or if it should
apply them as well.