Add general ToC and one to each subsection
Though the README contains great content, I found that I had to basically read the entire document to understand how to specifically use and configure Terragrunt. Adding tables of contents to each section will provide an overview of the contents of that section and allow users to easily skip to relevant/interesting sections without reading the entire document.