Rename helper function to built-in functions
The thought here is that users of Terragrunt will already be familiar with the terminology of [Terraform interpolation syntax](https://www.terraform.io/docs/configuration/interpolation.html) and built-in functions. The helper functions are really just Terragrunt built-in functions that are called using the same interpolation syntax as Terraform, so I tried to make the terminology consistent between the two tools.