Update description to add DRY and remove locking
Now that Terraform supports state locking natively, it appears that Terragrunt is getting out of the business of handling state locking, so I removed it from the tool description. The new use cases sections highlight the ability to keep your Terraform scripts DRY, so I added that to the short tool description too.