Create an Install Terragrunt section
On OSX, I usually try to install a tool via homebrew before anything else. The README didn't mention that was an option for Terragrunt, so I think highlighting that will make it even easier for people to install. The install section can be expanded for other methods in the future too. For example, maybe Terragrunt gets added to apt repos for Ubuntu, etc.