Clarify that include only copies Terragrunt config
As it was worded, it sounded to me like `include` would behave as if I had copy/pasted the *entire* terraform.tfvars file from the root to the child (e.g. including variables). This seemed odd to me, so I had to test it out to see the behavior first hand. I see that only the Terragrunt configuration is copied from the root to the child, so hopefully this small wording change will make that more clear.