Create separate migration guide file
Localizing the migration information into a separate guide will give
users a single document to review during their upgrade, as opposed
to reading multiple sections of the README file. The migration_guides
directory creates a place for all future migration guides to live. 

The README file is already quite large and this will help reduce
the size and scope of that file. The migration information is most
important to existing users of Terragrunt, but new users who are
starting from scratch may opt to skip that information since it is
legacy. Having the README clearly highlight and link to the migration
guide where appropriate means that users will still be able to easily
find upgrade information and choose whether or not it applies to them.