remote: Check whether input is a TTY before prompting
Add a new CanAcceptPrompt API for determining whether the user can
type in things to send to the running process. If the user cannot do
that, exit with an error immediately, rather than hanging forever.

I haven't checked whether this code is used in other parts of the
program, but if so we should probably update those as well.

Fixes #317.