Remove subdirectories before copying files to them
If a file in a subdirectory has permissions of 444, we will attempt to
open it for writing in util.CopyFile before overwriting it. This
fails because the file is read-only.

Instead remove all files in subdirectories. This shouldn't be much
slower than the current procedure, because we are still overwriting
every file every time, we're just starting from an empty file tree
instead of a full file tree. This also avoids the problem where run
1 produces file A, and run 2 does not have file A, but it exists in
the cache.

Fixes #337.