Added "run_on_error" logic.
This logic will allow the user to configure their before/after hooks to run or skip execution depending on whether a prior error has occurred.

For example: If an error occurs in a before hook - then we will not execute the terraform command and will only execute the after hooks which are configured to run even if an error has happened.