Removes Terragrunt's requirement for "//" in all module sources when using --terragrunt-source
This is accomplished by utilizing Hashicorp's go-getter to parse the module URLs.
The regexp was updated to also parse out the module name from each URL if the
module lives at the root of a repo. Test cases were updated to account for
HTTP paths, forward slashes in ?ref=, as well as local file sources.
The error check for a missing "//" was removed, as there are too many edge cases
based on a large amount of source formats supported by Terraform. Terraform itself
does not perform this validation on download.