Move bucket creation check
Terragrunt has a method that checks if someone is already creating an S3 bucket concurrently. However, even if that error is detected, after detecting it, we continued to try to configure the bucket in other ways (e.g., enable versioning) anyway. This PR updates the code to to skip all the other bucket creation steps if someone else is already doing them concurrently.