Adds options attribute to track state of the terraform command
The "init" hook command was previously overloaded, and would execute
hooks twice. Once when retrieving source using -from-module, and once
to configure the backend, modules, and plugins.

The new options attribute, `TerraformCommand`, is first set to the
terraform command when the `TerragruntOptions` are parsed from the
CLI arguments.

The `TerraformCommand` attribute is updated when `runTerraformInit`
determines that it needs to retrieve the source using `-from-module`.

Finally, the `TerraformCommand` is now used to determine whether a
hook should execute.

There are now two "init" commands that execute hooks:

* `init-from-module`: executes when the source is retrieved
* `init`: executes for any other init invocation, to configure
the backend, modules, and plugins

Fixes #539