Speed up init usage
When we use init to download modules, set `-get=false`, `-get-plugins=false`, and `-backend=false` so that all of those can be handled in the second call to `init` (if necessary). I tried this before in https://github.com/gruntwork-io/terragrunt/pull/516, but had to revert it due to a Terraform bug (https://github.com/hashicorp/terraform/issues/18460). I realize now that I can simply catch the error from that Terraform bug and ignore it!