Don't automatically apply external dependencies in non-interactive mode
The `xxx-all` commands were automatically selecting "yes" for all prompts if `--terragrunt-non-interactive` was set. This had the unfortunate side effect of automatically applying all commands to "external dependencies" (those outside the working dir), which is a very risky default when you run `destroy-all`. This change removes that default to be extra paranoid and never automatically delete external dependencies.

Note that this was exposed by in the `TestTerragruntStackCommands` test. It must have been passing before Terraform 0.12 due to 0.11.x swallowing errors?