Add CLI option to ignore the dependency order during *-all commands
When running `*-all` commands, we may want to run the modules while
ignoring the dependency order: during `plan-all` and `validate-all`,
since the state is not altered between each module execution,
we might as well run all the commands at once.

This adds an option, `IgnoreDependencyOrder` (defaulting to False),
which when set makes `*-all` disregard the dependency order.
This leverages on the `IgnoreOrder` run strategy introduced in the
previous commit.