Document Exclude-dir behaviour in ReadMe and tests
When using the `--terragrunt-exclude-dir`, modules are flagged
with a `FlagExcluded` boolean. Later on, when Terragrunt converts
the modules into runningModules (in the `toRunningModules` method),
it calls as last step the method `removeFlagExcluded`.

That method returns a modified map of `runningModules` map,
with the ones marked with the boolean removed and modules
which depend on them have their `Dependencies` fixed accordingly.

This adds unit tests for three use cases for `removeFlagExcluded`:
 - No module is marked with `FlagExcluded`:
   nothing happens, the output is the same as the input
 - One module, with no dependencies, is marked with `FlagExcluded`:
   that module is removed from the map
 - One module, with dependencies, is marked with `FlagExcluded`:
   that module is removed from the map and its direct depedency
   is altered to not depend on the excluded module anymore.