Improve docs by linking to DRY definition
Beginners and people less knowlegdable about industry jargon (like DRY)
might be confused when they encounter abbreviations without any
explanation to them at first use of said abbreviation.

It is good practice to introduce abbreviations with either writing them
out in full and putting the abbreviation in parentheses, or with
hypertext simply by making the abbreviation a link and pointing to the
full form.

The latter approach was chosen for this change.