Update find_in_parent_folders to return absolute path
Currently, the `find_in_parent_folders` function returns a relative path to the file it finds. This PR updates it to return the absolute path to that file. Depending on where you are using `find_in_parent_folders`, this can be important: e.g., one place I saw the relative path cause an issue is:

1. I had a child `terragrunt.hcl` that used `include` on a a parent `terragrunt.hcl`.
1. The parent `terragrunt.hcl` called `read_terragrunt_config(find_in_parent_folder("foo.hcl"))`
1. `foo.hcl` called `read_terragrunt_config(find_in_parent_folder("bar.hcl"))`
1. This led to an error, even though `bar.hcl` existed. I believe the error was because we were returning a relative path for it, but with `.terragrunt-cache/xxx/yyy` as the working dir, that ended up being the wrong path.

I'm not aware of any case where a relative path would have an advantage over an absolute path, so this is a backwards incompatible change to always return an absolute path.