Fix #1113: Makefile: add convenience 'build' target w/ dynamic version
New 'Makefile' targets include: 'build', 'terragrunt', and 'clean'; 'build' is
merely a convenience for those building from source.

The 'main.VERSION' value is here obtained dynamically from 'git-describe(1)'.
For end-users building from source out of the git repo, it allows the
'terragrunt' binary to be easily built with correct version information
without the user having to specify it manually (or even knowing how to do
so). When built in this way, the 'terragrunt --version' output will reflect
the git tag (and possibly additional commit info).