Deprecate "lock_table" attribute from use in the remote S3 backend configuration
The "lock_table" attribute of the remote S3 backend configuration has
been deprecated in Terraform before release 0.13 where it has been
retired. The current name of the attribute is "dynamodb_table".

Terragrunt currently supports both of the attributes, but should the
"lock_table" one be used, then it would be passed as-is either when
generating a remote S3 backend block when templating Terraform or as
a command-line argument. Using older attribute would not be a problem
with Terraform before release 0.13, where the old attribute is still
present albeit deprecated, but for anyone who uses the latest release,
this would now become a problem as Terraform would terminate with an
error.

This change deprecates the use of the "lock_table" in Terragrunt for
remote S3 backend configuration and adds facilities to notify the user
about the attribute being deprecated. Additionally, the use of
"lock_table" will be normalised so that the "dynamodb_table" will be
used instead in the resulting remote S3 backend block or as
a command-line argument.

Signed-off-by: Krzysztof Wilczyński <kw@linux.com>