Make Terragrunt recover from missing module and provider error
This change enables Terragrunt to detect errors that otherwise required
the user to run "terraform init" and attempt a recovery by running
Terraform automatically.

When a Terraform module or provider version changes or when a module or
provider is missing and Terragrunt is invoked with an action other than
"init" (e.g., plan, apply, etc.), then the Terraform run that has been
started by Terragrunt will terminate with an error asking the user to
run "terraform init" to reinitialize to fetch the missing runtime
dependencies.

Terragrunt should detect these type of missing module and provider
errors and automatically re-run "terraform init" to resolve them and
retry the previous command automatically.

Signed-off-by: Krzysztof Wilczyński <kw@linux.com>