Update Terragrunt to work with Terraform 0.14 (#1459)
* WIP: Add support for terraform 0.14 lock files

* Copying lock file

* returning new argument to match functiona signature

* Updating additional function with originalTerragruntOptions

* running terraform fmt

* Updating local download test to check for TF lock file

* Adding custom lock file test + fixture

* Finish test for custom lock file

* Fix all test failures due to terraform output changes

See https://github.com/hashicorp/terraform/issues/27100 for context

* Fix issue with Terragrunt always calling init

Fixes #1423. Terraform 0.14 stores the data in a providers folder rather than plugins.

* Change custom-lock-file test to use separate module

It was originally using the "hello-world" module, but that module is used in other tests, and those tests may run 'init' in it, creating a different lock file than the one we're looking for in our test.

* Fix TestTerragruntStdOut

We need newlines, not literal '\n'

* Fix source URL in TestCustomLockFile

Also, add/update comments to clarify what it is we're testing

* Fix TestCustomLockFile

We can't rely on a comment in the lock file, as Terraform overwrites it every time. But we can rely on the versions being preserved!

* Add docs on how Terragrunt handles lock files

* Add missing words to comments

* Add comment on why we use working dir for lock file

* Fix lock file name

Should be .terraform.lock.hcl and not just .terraform.lock

Co-authored-by: Ina Stoyanova <ina@gruntwork.io>
Co-authored-by: Yevgeniy Brikman <brikis98@gmail.com>