Prefer remote tfstate over local tfstate, if present (#1885)
This addresses issue #514:

Upon a local to remote tfstate migration, Terraform will leave an empty
`terraform.tfstate` file in the working directory. Subsequent
invocations on the state should then refer to the `terraform.tfstate`
file in the data directory (`.terraform`) which points to the remote
state. By prefering the data directory when parsing the tfstate,
Terragrunt behaves similarly to Terraform, and no longer errors out on
the empty `terraform.tfstate` file.

Co-authored-by: Andrey Kaipov <andreykaipov@users.noreply.github.com>