Use faster hashing approach when handling local module sources
The current approach to hashing (sha256 of the file contents of each
module) is very slow on large module trees due to a combination of doing
lots of File I/O and spending lots of CPU time on running sha256
calculations.

This manifests as Terragrunt spending 5+ minutes at 100% CPU at the
beginning of every run, with no output.

This commit takes a faster, if slightly less perfect, approach to
determining if a local module has had its files modified. Rather than
hashing the entire contents of each file, it takes the `mtime` of the
file (obtained through a cheap `stat` call), appends that to the file
name, and hashes that combined string.

The downside of this approach is that if a file has been written to
since we last copied the module, we'll copy it again. Thankfully, that
is exceedingly unlikely to happen in the middle of a Terragrunt run.