fix!: output-module-groups to work like run-all
output-module-groups was not working from a directory that did not have
a terragrunt.hcl resource and it was not working with terragrunt cli options
such as `terragrunt-modules-that-include`.

furthermore, it was not inverting order for the destroy command.

and required init even though it is just using the stackconfig to output
planned behavior.

this change aligns the behavior with run-all so that it will work with
terragrunt cli options, properly display order based on apply and/or destory,
and can be run withing any folder that run-all could be executed in.

resolves: #2685
resolves: #2768

BREAKING CHANGE: technically this does change the behavior, in that it no longer requires terraform init. but i believe that more accurately reflects the desired behavior.