fix(provider_cache): Use more permissive regex search for Status Locked
When 2 following conditions are met, then terragrunt can't find 423 Locked in the output:

* `--terragrunt-include-module-prefix` is enabled
* Provider name / version is long enough

The output from terraform looks like this:

```
╷
│ Error: Failed to install provider
│
│ Error while installing snowflake-labs/snowflake v0.89.0: could not query
│ provider registry for registry.terraform.io/snowflake-labs/snowflake: 423
│ Locked
╵
```

And the regex which is used right now (`(?mi)423[^a-z0-9]*Locked`)
avoids non-alphanumeric characters that are prepended by the
IncludeModulePrefix.

Let's use more relaxed regex to match even prefixed otuput.

Fixes: https://github.com/gruntwork-io/terragrunt/issues/3100