`sops_decrypt_file()`: resolve path argument relative to `terragrunt.hcl` (#2752)
* Fix relative paths with sops_decrypt_file()

sops_decrypt_file() computed the file path relative to the working
directory instead of relative to the Terragrunt configuration. This
generally worked with individual `terragrunt apply` operations but
failed for `terragrunt run-all apply`, as it would look for the file in
the parent directory.

Fixes #1319.

* Add test case for sops_decrypt_file() + run-all

Verify that sops_decrypt_file() behaves as expected in combination with
run-all.