refactor: replace `golang.org/x/exp` with stdlib
The experimental functions are now available in the standard library in
Go 1.21 [1] and Go 1.23 [2].

The difference is that `maps.Keys`/`maps.Values` in the
`golang.org/x/exp/maps` package return a slice, whereas
`maps.Keys`/`maps.Values` in the standard library return an iterator. To
work with slices, we need to use `slices.Collect` to convert the
iterator into a slice.

[1]: https://go.dev/doc/go1.21#slices
[2]: https://go.dev/doc/go1.23#iterators

Signed-off-by: Eng Zer Jun <engzerjun@gmail.com>