A collection of website polishing (#4784)
* Tighten up sidebar.

* Standardize CSS ordering between dev and prod.

Previously, the site would render one way in prod and another in dev! The issue was that vite was tree-shaking CSS and wound up re-ordering things in a way it thought was permissible, whereas in dev, without that optimization, the CSS was actuallys sequenced differently.

This led to a noticable huge margin in dev, but not in prod.

This commit asserts the official ordering the CSS layers, and then adds a fix for the left-margin issue. This should now standard dev and prod.

* Reduce code font-size from 16px to 14px.

* Improve main paragraph text rendering.

Use more readable line height and paragraph separation.

* Improve spacing after file tree.

* Fix sidebar inconsistencies on 3rd level of nav.